added timestamp to db when track was last played

+++ createMusicLib.py
@@ -40,7 +40,7 @@ cursor.execute("""DROP TABLE IF EXISTS musiclib """)
 # Pfad / Artist / Title / Album / Tracklaenge / Votes / LastTimePlayed in seconds
 
 cursor.execute("""CREATE TABLE IF NOT EXISTS musiclib ( 
-    		id INTEGER, path TEXT, artist TEXT, title TEXT, album TEXT, albumart TEXT, year TEXT, tracklength INTEGER , votes INTEGER, lastplayed INTEGER)""")
+    		id INTEGER, path TEXT, artist TEXT, title TEXT, album TEXT, albumart TEXT, year TEXT, tracklength INTEGER , votes INTEGER, lastplayed FLOAT)""")
 
 # add every musicfile with id3tag information to db
 id = 0
